refactor: Update import paths for ChatContext and related components, reorganize file structure
