+++ docs/NAMECHEAP-MIGRATION.md
@@ -1,0 +1,49 @@
+# SnakkaZ Chat Migration from Cloudflare to Namecheap
+
+This document outlines the migration process from Cloudflare dependencies to Namecheap for the SnakkaZ Chat application.
+
+## Changes Made
+
+### 1. File Reorganization
+Files have been moved from `/services/encryption` to more logical locations:
+
+- UI Components → `/components/chat/`
+- Contexts → `/contexts/`
+- Encryption Utilities → `/utils/encryption/`
+- API Services → `/services/api/`
+- Theme Service → `/utils/`
+- DNS Services → `/services/dns/`
+- Scripts → `/scripts/`
+
+### 2. Cloudflare Dependency Removal
+The following Cloudflare-specific files/code have been removed:
+
+- Cloudflare Analytics in `analyticsLoader.ts`
+- Cloudflare domains from CSP in `snakkazCspPlugin.ts`
+- Cloudflare Analytics script injection from `inject-csp.sh`
+- Cloudflare-specific reporting in `cspReporting.ts`
+- Various Cloudflare helper functions and fixes
+
+### 3. CSP Configuration Updates
+Content Security Policy has been updated to work with Namecheap instead of Cloudflare:
+
+- Removed Cloudflare-specific domains from CSP directives
+- Updated CSP reporting endpoints
+- Simplified CSP configuration
+
+### 4. Import Path Updates
+All files that referenced the old structure have been updated to use the new import paths.
+
+## Migration Process
+
+1. Created backup of original files
+2. Reorganized file structure into logical categories
+3. Updated all import paths
+4. Removed Cloudflare-specific code
+5. Tested application functionality
+
+## Remaining Tasks
+
+- Configure DNS with Namecheap API
+- Set up proper CSP reporting without Cloudflare
+- Update any documentation referencing Cloudflare
