+++ src/components/chat/ChatInterface.tsx
@@ -6,7 +6,7 @@
  */
 
 import React, { useState, useRef, useEffect } from 'react';
-import { useChat } from '../../services/encryption/ChatContext';
+import { useChat } from '../../contexts/ChatContext';
 import { Button } from '../../components/ui/button';
 import { Input } from '../../components/ui/input';
 import { Loader2, Send, Image, Paperclip, Smile, X, Check, CheckCheck } from 'lucide-react';
